+++ common/src/main/java/com/st0x0ef/stellaris/client/overlays/RocketBarOverlay.java
@@ -5,6 +5,7 @@ import com.st0x0ef.stellaris.client.screens.helper.ScreenHelper;
 import com.st0x0ef.stellaris.common.entities.vehicles.LanderEntity;
 import com.st0x0ef.stellaris.common.entities.vehicles.RocketEntity;
 import com.st0x0ef.stellaris.common.utils.PlanetUtil;
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
@@ -14,11 +15,9 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 
-import static com.st0x0ef.stellaris.Stellaris.texture;
-
 public class RocketBarOverlay {
 
-    public static final ResourceLocation ROCKET = texture("planet_bar/rocket");
+    public static final ResourceLocation ROCKET = ResourceLocationUtils.texture("planet_bar/rocket");
 
     public static void render(GuiGraphics graphics, DeltaTracker deltaTracker) {
         Player player = Minecraft.getInstance().player;