+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/MilkyWayScreen.java
@@ -3,6 +3,7 @@ package com.st0x0ef.stellaris.client.screens;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.st0x0ef.stellaris.client.screens.helper.ScreenHelper;
 import com.st0x0ef.stellaris.common.menus.MilkyWayMenu;
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
@@ -15,14 +16,11 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 import org.lwjgl.glfw.GLFW;
 
-import static com.st0x0ef.stellaris.Stellaris.guiTexture;
-import static com.st0x0ef.stellaris.Stellaris.texture;
-
 @Environment(EnvType.CLIENT)
 public class MilkyWayScreen extends AbstractContainerScreen<MilkyWayMenu> {
 
-    public static final ResourceLocation MILKY_WAY_TEXTURE = texture("environment/milky_way");
-    public static final ResourceLocation BACKGROUND_TEXTURE = guiTexture("planet_selection");
+    public static final ResourceLocation MILKY_WAY_TEXTURE = ResourceLocationUtils.texture("environment/milky_way");
+    public static final ResourceLocation BACKGROUND_TEXTURE = ResourceLocationUtils.guiTexture("planet_selection");
 
     public static float rotationAngle = 0;