+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/RadioactiveGeneratorScreen.java
@@ -4,6 +4,7 @@ import com.mojang.blaze3d.systems.RenderSystem;
 import com.st0x0ef.stellaris.client.screens.components.GaugeWidget;
 import com.st0x0ef.stellaris.common.blocks.entities.machines.RadioactiveGeneratorEntity;
 import com.st0x0ef.stellaris.common.menus.RadioactiveGeneratorMenu;
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
@@ -15,12 +16,10 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Inventory;
 
-import static com.st0x0ef.stellaris.Stellaris.guiTexture;
-
 @Environment(EnvType.CLIENT)
 public class RadioactiveGeneratorScreen extends AbstractContainerScreen<RadioactiveGeneratorMenu> {
 
-    public static final ResourceLocation TEXTURE = guiTexture("radioactive_generator");
+    public static final ResourceLocation TEXTURE = ResourceLocationUtils.guiTexture("radioactive_generator");
 
     private final RadioactiveGeneratorEntity blockEntity = getMenu().getBlockEntity();
     private GaugeWidget energyGauge;