+++ common/src/main/java/com/st0x0ef/stellaris/common/blocks/entities/machines/CoalGeneratorEntity.java
@@ -99,7 +99,6 @@ public class CoalGeneratorEntity extends BaseGeneratorBlockEntity {
             energyContainer.insert(energyGeneratedPT, false);
         }
 
-        EnergyUtil.moveEnergyToItem(getEnergy(null), getItem(1), 10);
         EnergyUtil.distributeEnergyNearby(level, worldPosition, maxCapacity);
     }
 
@@ -140,7 +139,6 @@ public class CoalGeneratorEntity extends BaseGeneratorBlockEntity {
 
     @Override
     public int getContainerSize() {
-        return 2;
+        return 1;
     }
-
 }