+++ common/src/main/java/com/st0x0ef/stellaris/common/network/packets/SyncEnergyPacketWithoutDirection.java
@@ -2,6 +2,7 @@ package com.st0x0ef.stellaris.common.network.packets;
 
 import com.fej1fun.potentials.energy.BaseEnergyStorage;
 import com.fej1fun.potentials.providers.EnergyProvider;
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import dev.architectury.networking.NetworkManager;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.client.Minecraft;
@@ -12,11 +13,9 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import org.jetbrains.annotations.NotNull;
 
-import static com.st0x0ef.stellaris.Stellaris.id;
-
 public record SyncEnergyPacketWithoutDirection(int energy, BlockPos pos) implements CustomPacketPayload {
 
-    public static final Type<SyncEnergyPacketWithoutDirection> TYPE = new Type<>(id("energy_sync_packet_without_direction"));
+    public static final Type<SyncEnergyPacketWithoutDirection> TYPE = new Type<>(ResourceLocationUtils.id("energy_sync_packet_without_direction"));
     public static final StreamCodec<ByteBuf, SyncEnergyPacketWithoutDirection> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.VAR_INT, SyncEnergyPacketWithoutDirection::energy,
             BlockPos.STREAM_CODEC, SyncEnergyPacketWithoutDirection::pos,