+++ common/src/main/java/com/st0x0ef/stellaris/common/network/packets/SyncFluidPacketWithoutDirection.java
@@ -3,6 +3,7 @@ package com.st0x0ef.stellaris.common.network.packets;
 import com.fej1fun.potentials.components.FluidAmountMapDataComponent;
 import com.fej1fun.potentials.fluid.BaseFluidStorage;
 import com.fej1fun.potentials.providers.FluidProvider;
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import com.st0x0ef.stellaris.common.utils.capabilities.fluid.SingleFluidStorage;
 import dev.architectury.fluid.FluidStack;
 import dev.architectury.networking.NetworkManager;
@@ -15,12 +16,10 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import org.jetbrains.annotations.NotNull;
 
-import static com.st0x0ef.stellaris.Stellaris.id;
-
 public record SyncFluidPacketWithoutDirection(FluidAmountMapDataComponent fluid, int tank,
                                               BlockPos pos) implements CustomPacketPayload {
 
-    public static final Type<SyncFluidPacketWithoutDirection> TYPE = new Type<>(id("fluid_sync_packet_without_direction"));
+    public static final Type<SyncFluidPacketWithoutDirection> TYPE = new Type<>(ResourceLocationUtils.id("fluid_sync_packet_without_direction"));
     public static final StreamCodec<RegistryFriendlyByteBuf, SyncFluidPacketWithoutDirection> STREAM_CODEC = StreamCodec.composite(
             FluidAmountMapDataComponent.STREAM_CODEC, SyncFluidPacketWithoutDirection::fluid,
             ByteBufCodecs.VAR_INT, SyncFluidPacketWithoutDirection::tank,