+++ common/src/main/java/com/st0x0ef/stellaris/common/network/packets/SyncPlanetMenuState.java
@@ -1,5 +1,6 @@
 package com.st0x0ef.stellaris.common.network.packets;
 
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import dev.architectury.networking.NetworkManager;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
@@ -7,13 +8,11 @@ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.world.entity.player.Player;
 import org.jetbrains.annotations.NotNull;
 
-import static com.st0x0ef.stellaris.Stellaris.id;
-
 public class SyncPlanetMenuState implements CustomPacketPayload {
 
     final boolean open;
 
-    public static final Type<SyncPlanetMenuState> TYPE = new Type<>(id("sync_planet_menu_state"));
+    public static final Type<SyncPlanetMenuState> TYPE = new Type<>(ResourceLocationUtils.id("sync_planet_menu_state"));
 
     public static final StreamCodec<RegistryFriendlyByteBuf, SyncPlanetMenuState> STREAM_CODEC = new StreamCodec<>() {