+++ common/src/main/java/com/st0x0ef/stellaris/common/registry/EntityRegistry.java
@@ -15,6 +15,7 @@ import com.st0x0ef.stellaris.common.entities.mobs.pygro.PygroMobsSensor;
 import com.st0x0ef.stellaris.common.entities.vehicles.LanderEntity;
 import com.st0x0ef.stellaris.common.entities.vehicles.RocketEntity;
 import com.st0x0ef.stellaris.common.entities.vehicles.RoverEntity;
+import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;
 import dev.architectury.registry.level.entity.SpawnPlacementsRegistry;
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
@@ -30,52 +31,50 @@ import net.minecraft.world.level.levelgen.Heightmap;
 import java.util.function.BiConsumer;
 import java.util.function.Supplier;
 
-import static com.st0x0ef.stellaris.Stellaris.id;
-
 //For everything related to entities
 public class EntityRegistry {
     //Entity type
 
     public static final DeferredRegister<EntityType<?>> ENTITY_TYPE = DeferredRegister.create(Stellaris.MODID, Registries.ENTITY_TYPE);
     public static final RegistrySupplier<EntityType<Alien>> ALIEN = ENTITY_TYPE.register("alien",
-            Suppliers.memoize(() -> EntityType.Builder.of(Alien::new, MobCategory.MONSTER).sized(0.75f, 2.5f).build(id("alien").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(Alien::new, MobCategory.MONSTER).sized(0.75f, 2.5f).build(ResourceLocationUtils.id("alien").toString())));
     public static final RegistrySupplier<EntityType<AlienZombie>> ALIEN_ZOMBIE = ENTITY_TYPE.register("alien_zombie",
-            Suppliers.memoize(() -> EntityType.Builder.of(AlienZombie::new, MobCategory.MONSTER).sized(0.6f, 2.4f).build(id("alien_zombie").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(AlienZombie::new, MobCategory.MONSTER).sized(0.6f, 2.4f).build(ResourceLocationUtils.id("alien_zombie").toString())));
     public static final RegistrySupplier<EntityType<MartianRaptor>> MARTIAN_RAPTOR = ENTITY_TYPE.register("martian_raptor",
-            Suppliers.memoize(() -> EntityType.Builder.of(MartianRaptor::new, MobCategory.MONSTER).sized(0.75f, 2.0f).build(id("martian_raptor").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(MartianRaptor::new, MobCategory.MONSTER).sized(0.75f, 2.0f).build(ResourceLocationUtils.id("martian_raptor").toString())));
     public static final RegistrySupplier<EntityType<Pygro>> PYGRO = ENTITY_TYPE.register("pygro",
-            Suppliers.memoize(() -> EntityType.Builder.of(Pygro::new, MobCategory.MONSTER).sized(0.6f, 1.8f).build(id("pygro").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(Pygro::new, MobCategory.MONSTER).sized(0.6f, 1.8f).build(ResourceLocationUtils.id("pygro").toString())));
     public static final RegistrySupplier<EntityType<PygroBrute>> PYGRO_BRUTE = ENTITY_TYPE.register("pygro_brute",
-            Suppliers.memoize(() -> EntityType.Builder.of(PygroBrute::new, MobCategory.MONSTER).sized(0.6f, 1.8f).build(id("pygro_brute").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(PygroBrute::new, MobCategory.MONSTER).sized(0.6f, 1.8f).build(ResourceLocationUtils.id("pygro_brute").toString())));
     public static final RegistrySupplier<EntityType<Mogler>> MOGLER = ENTITY_TYPE.register("mogler",
-            Suppliers.memoize(() -> EntityType.Builder.of(Mogler::new, MobCategory.MONSTER).sized(1.4f, 1.4f).build(id("mogler").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(Mogler::new, MobCategory.MONSTER).sized(1.4f, 1.4f).build(ResourceLocationUtils.id("mogler").toString())));
     public static final RegistrySupplier<EntityType<StarCrawler>> STAR_CRAWLER = ENTITY_TYPE.register("star_crawler",
-            Suppliers.memoize(() -> EntityType.Builder.of(StarCrawler::new, MobCategory.MONSTER).sized(1.3f, 1.0f).build(id("star_crawler").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(StarCrawler::new, MobCategory.MONSTER).sized(1.3f, 1.0f).build(ResourceLocationUtils.id("star_crawler").toString())));
     public static final RegistrySupplier<EntityType<? extends IceSpit>> ICE_SPIT = ENTITY_TYPE.register("ice_spit",
-            Suppliers.memoize(() -> EntityType.Builder.of(IceSpit::new, MobCategory.MISC).sized(0.5f, 0.5f).build(id("ice_spit").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(IceSpit::new, MobCategory.MISC).sized(0.5f, 0.5f).build(ResourceLocationUtils.id("ice_spit").toString())));
     public static final RegistrySupplier<EntityType<IceShardArrowEntity>> ICE_SHARD_ARROW = ENTITY_TYPE.register("ice_shard_arrow",
-            Suppliers.memoize(() -> EntityType.Builder.<IceShardArrowEntity>of(IceShardArrowEntity::new, MobCategory.MISC).sized(0.5f, 0.5f).build(id("ice_shard_arrow").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.<IceShardArrowEntity>of(IceShardArrowEntity::new, MobCategory.MISC).sized(0.5f, 0.5f).build(ResourceLocationUtils.id("ice_shard_arrow").toString())));
     public static final RegistrySupplier<EntityType<CheeseBoss>> CHEESE_BOSS = ENTITY_TYPE.register("cheese_boss",
-            Suppliers.memoize(() -> EntityType.Builder.of(CheeseBoss::new, MobCategory.MONSTER).sized(1.0f, 3.7f).eyeHeight(3.5f).build(id("cheese_boss").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(CheeseBoss::new, MobCategory.MONSTER).sized(1.0f, 3.7f).eyeHeight(3.5f).build(ResourceLocationUtils.id("cheese_boss").toString())));
     public static final RegistrySupplier<EntityType<? extends CheeseSpit>> CHEESE_SPIT = ENTITY_TYPE.register("cheese_spit",
-            Suppliers.memoize(() -> EntityType.Builder.of(CheeseSpit::new, MobCategory.MISC).sized(0.5f, 0.5f).build(id("cheese_spit").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(CheeseSpit::new, MobCategory.MISC).sized(0.5f, 0.5f).build(ResourceLocationUtils.id("cheese_spit").toString())));
 
     public static final RegistrySupplier<EntityType<CustomLightningBolt>> CUSTOM_LIGHTNING_BOLT = ENTITY_TYPE.register("custom_lightning_bolt",
-            Suppliers.memoize(() -> EntityType.Builder.of(CustomLightningBolt::new, MobCategory.MISC).sized(0.5f, 0.5f).build(id("custom_lightning_bolt").toString())));
+            Suppliers.memoize(() -> EntityType.Builder.of(CustomLightningBolt::new, MobCategory.MISC).sized(0.5f, 0.5f).build(ResourceLocationUtils.id("custom_lightning_bolt").toString())));
 
 
     /**
      * Vehicles
      */
 
-    public static final RegistrySupplier<EntityType<RocketEntity>> TINY_ROCKET = ENTITY_TYPE.register("tiny_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(1.1f, 4.4f).fireImmune().build(id("tiny_rocket").toString()));
-    public static final RegistrySupplier<EntityType<RocketEntity>> SMALL_ROCKET = ENTITY_TYPE.register("small_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(0.8f, 3.2f).build(id("small_rocket").toString()));
-    public static final RegistrySupplier<EntityType<RocketEntity>> NORMAL_ROCKET = ENTITY_TYPE.register("normal_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(1.1f, 4.4f).build(id("normal_rocket").toString()));
-    public static final RegistrySupplier<EntityType<RocketEntity>> BIG_ROCKET = ENTITY_TYPE.register("big_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(1.1f, 4.4f).fireImmune().build(id("big_rocket").toString()));
+    public static final RegistrySupplier<EntityType<RocketEntity>> TINY_ROCKET = ENTITY_TYPE.register("tiny_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(1.1f, 4.4f).fireImmune().build(ResourceLocationUtils.id("tiny_rocket").toString()));
+    public static final RegistrySupplier<EntityType<RocketEntity>> SMALL_ROCKET = ENTITY_TYPE.register("small_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(0.8f, 3.2f).build(ResourceLocationUtils.id("small_rocket").toString()));
+    public static final RegistrySupplier<EntityType<RocketEntity>> NORMAL_ROCKET = ENTITY_TYPE.register("normal_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(1.1f, 4.4f).build(ResourceLocationUtils.id("normal_rocket").toString()));
+    public static final RegistrySupplier<EntityType<RocketEntity>> BIG_ROCKET = ENTITY_TYPE.register("big_rocket", () -> EntityType.Builder.of(RocketEntity::new, MobCategory.MISC).sized(1.1f, 4.4f).fireImmune().build(ResourceLocationUtils.id("big_rocket").toString()));
 
-    public static final RegistrySupplier<EntityType<RoverEntity>> ROVER = ENTITY_TYPE.register("rover", () -> EntityType.Builder.of(RoverEntity::new, MobCategory.MISC).sized(2.7F, 1.3F).fireImmune().build(id("rover").toString()));
+    public static final RegistrySupplier<EntityType<RoverEntity>> ROVER = ENTITY_TYPE.register("rover", () -> EntityType.Builder.of(RoverEntity::new, MobCategory.MISC).sized(2.7F, 1.3F).fireImmune().build(ResourceLocationUtils.id("rover").toString()));
 
-    public static final RegistrySupplier<EntityType<LanderEntity>> LANDER = ENTITY_TYPE.register("lander", () -> EntityType.Builder.<LanderEntity>of(LanderEntity::new, MobCategory.MISC).sized(2.5f, 1.0f).build(id("lander").toString()));
+    public static final RegistrySupplier<EntityType<LanderEntity>> LANDER = ENTITY_TYPE.register("lander", () -> EntityType.Builder.<LanderEntity>of(LanderEntity::new, MobCategory.MISC).sized(2.5f, 1.0f).build(ResourceLocationUtils.id("lander").toString()));
 
     public static void registerAttributes(BiConsumer<Supplier<? extends EntityType<? extends LivingEntity>>, Supplier<AttributeSupplier.Builder>> attributes) {
         attributes.accept(ALIEN, Alien::setCustomAttributes);