+++ changelog.md
@@ -10,6 +10,8 @@
 - Jet suit and space suit now need diesel instead of fuel
 - Add emoji symbols to tooltips
 - Oil finder now take energy
+- Update almost all gui style
+- Change some machines' max energy capacity
 
 ## Fixes
 - Space suit crash when trying to put fuel in it